Social shares for proposals
