+++ components/proposals/ProposalPage/components/ProposalEvaluations/components/Review/EvaluationsReview.tsx
@@ -14,9 +14,9 @@ import { EvaluationStepRow } from './components/EvaluationStepRow';
 import { EvaluationStepSettingsModal } from './components/EvaluationStepSettingsModal';
 import { FeedbackEvaluation } from './components/FeedbackEvaluation';
 import { PassFailEvaluation } from './components/PassFailEvaluation';
-import { ProposalSocialShare } from './components/ProposalSocialShare';
 import { PublishRewardsButton } from './components/PublishRewardsButton';
 import { RubricEvaluation } from './components/RubricEvaluation/RubricEvaluation';
+import { ProposalSocialShareLinks } from './components/SocialShare/ProposalSocialShareLinks';
 import { VoteEvaluation } from './components/VoteEvaluation/VoteEvaluation';
 
 export type Props = {
@@ -224,7 +224,7 @@ export function EvaluationsReview({
       {pagePath && pageTitle && proposal && expandedContainer && (
         <>
           <Divider />
-          <ProposalSocialShare
+          <ProposalSocialShareLinks
             lensPostLink={proposal.lensPostLink}
             proposalId={proposal.id}
             proposalPath={pagePath}
